requirements.txt
# =============================================================================
# ENTERPRISE SECURITY MONITORING DASHBOARD - DEPENDENCIES
# =============================================================================

# Core Web Dashboard
dash>=2.14.0
plotly>=5.17.0
dash-bootstrap-components>=1.5.0
flask>=2.3.0
werkzeug>=2.3.0

# Data Processing & Analysis
pandas>=2.1.0
numpy>=1.25.0
python-dateutil>=2.8.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL support
pymysql>=1.1.0          # MySQL support

# Threat Intelligence APIs
requests>=2.31.0
urllib3>=2.0.0
certifi>=2023.7.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Security & Cryptography  
cryptography>=42.0.0

# System Monitoring
psutil>=5.9.0

# Task Scheduling
schedule>=1.2.0

# Data Validation
jsonschema>=4.20.0

# Testing Framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
flake8>=6.1.0

# Demo Data Generation
faker>=19.6.0

# Production Deployment (Optional)
gunicorn>=21.2.0        # WSGI server
redis>=5.0.0            # Caching (optional)
structlog>=23.1.0       # Structured logging

# =============================================================================
# INSTALLATION INSTRUCTIONS
# =============================================================================
# 
# Basic installation:
#   pip install -r requirements.txt
#
# Development installation:
#   pip install -r requirements.txt
#   pip install -e .
#
# Production installation:
#   pip install -r requirements.txt --no-dev
#
# =============================================================================